Split annotation
